MR#5930: Add the ability to adding interactive 2D widgets
  - avoid a warning

+++ Common/Core/vtkFFT.cxx
@@ -23,7 +23,8 @@ vtkStandardNewMacro(vtkFFT);
 //------------------------------------------------------------------------------
 std::vector<vtkFFT::ComplexNumber> vtkFFT::FftDirect(const std::vector<ScalarNumber>& in)
 {
-  kiss_fftr_cfg cfg = kiss_fftr_alloc(in.size(), 0 /*is_inverse_fft*/, nullptr, nullptr);
+  kiss_fftr_cfg cfg =
+    kiss_fftr_alloc(static_cast<int>(in.size()), 0 /*is_inverse_fft*/, nullptr, nullptr);
   if (cfg != nullptr)
   {
     auto outSize = (in.size() / 2) + 1;
@@ -41,7 +42,8 @@ std::vector<vtkFFT::ComplexNumber> vtkFFT::FftDirect(const std::vector<ScalarNum
 std::vector<vtkFFT::ScalarNumber> vtkFFT::FftInverse(const std::vector<vtkFFT::ComplexNumber>& in)
 {
   auto outSize = (in.size() - 1) * 2;
-  kiss_fftr_cfg cfg = kiss_fftr_alloc(outSize, 1 /*is_inverse_fft*/, nullptr, nullptr);
+  kiss_fftr_cfg cfg =
+    kiss_fftr_alloc(static_cast<int>(outSize), 1 /*is_inverse_fft*/, nullptr, nullptr);
   if (cfg != nullptr)
   {
     std::vector<vtkFFT::ScalarNumber> result(outSize);
